// src/app/components/auth/model/login-data.ts
import { LoggedinUser } from "../../../models/etesteruser";

export interface LoginDialogData {
  username: string;
  password: string;
  email: string;
  authToken: string;
  loggedinUser?: LoggedinUser;
}

export interface LoginEvent {
  username: string;
  email: string;
  authToken: string;
  loggedinUser?: LoggedinUser;
  isLoggedIn?: boolean;
}